Base folder can be specified in config file
A user can create a file ~/.config/ptlogger2/config.properties; if it
exists and has a property called `basefolder`, this folder is assumed to
be the location of log files for every file that is not preceded by an
absolute path. This way, the program can be called from any folder.